refactor(core): extract shared verified-sqlite helpers into sqlite-artifact module
